requirements.txt
requests>=2.28
playwright>=1.40
pyyaml>=6.0
